pyproject.toml
[project]
name = "tinkerforge2mqtt"
dynamic = ["version"]
description = "Emit MQTT events from Tinkerforge devices"
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
authors = [
    {name = 'Jens Diemer', email = 'git@jensdiemer.de'}
]
requires-python = ">=3.11"
dependencies = [
    "tinkerforge", # https://www.tinkerforge.com/doc/Source_Code.html
    "ha-services @ git+https://github.com/SuiOni/ha-services.git@0c7fb7d",
    "cli-base-utilities", # https://github.com/jedie/cli-base-utilities
    "bx_py_utils", # https://github.com/boxine/bx_py_utils
    "tyro", # https://github.com/brentyi/tyro
    "rich", # https://github.com/Textualize/rich
]

[dependency-groups]
dev = [
    "manageprojects",  # https://github.com/jedie/manageprojects
    "urllib3", # for bx_py_utils.test_utils.deny_requests.deny_any_real_request() in tests
    "uv",  # https://github.com/astral-sh/uv
    "hatchling",  # https://github.com/pypa/hatch/tree/master/backend
    "nox",  # https://github.com/wntrblm/nox
    "coverage",  # https://github.com/nedbat/coveragepy
    "ruff", # https://github.com/astral-sh/ruff
    "codespell",  # https://github.com/codespell-project/codespell
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
    "pip-audit",  # https://github.com/pypa/pip-audit
    "mypy",  # https://github.com/python/mypy
    "twine",  # https://github.com/pypa/twine
    "pre-commit",  # https://github.com/pre-commit/pre-commit
    "typeguard",  # https://github.com/agronholm/typeguard/
]

[project.urls]
Documentation = "https://github.com/jedie/tinkerforge2mqtt"
Source = "https://github.com/jedie/tinkerforge2mqtt"

[project.scripts]
tinkerforge2mqtt_app = "tinkerforge2mqtt.__main__:main"
tinkerforge2mqtt_dev = "tinkerforge2mqtt.cli_dev:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tinkerforge2mqtt"]

[tool.hatch.version]
path = "tinkerforge2mqtt/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.cli_base.pip_audit]
strict=true
require_hashes=true
ignore-vuln=[]

[tool.cli_base]
version_module_name = "tinkerforge2mqtt"  # Used by "update-readme-history" git hook


[tool.ruff]
# https://docs.astral.sh/ruff/configuration/
line-length = 120
exclude = [
  ".*/",
]

[tool.ruff.lint]
preview = true  # Needed for some of the rules
extend-select = [
    "E", # pycodestyle: https://docs.astral.sh/ruff/rules/#error-e
    "I", # isort: https://docs.astral.sh/ruff/rules/#isort-i
]
ignore = [
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
    "E731", # https://docs.astral.sh/ruff/rules/lambda-assignment/
]

[tool.ruff.lint.isort]
# https://docs.astral.sh/ruff/settings/#lintisort
lines-after-imports = 2

[tool.ruff.format]
quote-style = "single"


[tool.coverage.run]
branch = true
parallel = true
concurrency = ["multiprocessing"]
source = ['.']
command_line = '-m unittest --verbose --locals --buffer'

[tool.coverage.report]
omit = ['.*', '*/tests/*']
skip_empty = true
fail_under = 30
show_missing = true
exclude_lines = [
    'if self.debug:',
    'pragma: no cover',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
]


[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
allow_redefinition = true  # https://github.com/python/mypy/issues/7165
show_error_codes = true
plugins = []
exclude = ['.venv', 'tests']


[manageprojects] # https://github.com/jedie/manageprojects
initial_revision = "9e29a86"
initial_date = 2024-02-29T22:32:49+01:00
cookiecutter_template = "https://github.com/jedie/cookiecutter_templates/"
cookiecutter_directory = "uv-python"
applied_migrations = [
    "765b8fb", # 2024-03-12T09:08:01+01:00
    "2d314b4", # 2025-10-21T13:55:34+02:00
]

[manageprojects.cookiecutter_context.cookiecutter]
full_name = "Jens Diemer"
github_username = "jedie"
author_email = "git@jensdiemer.de"
package_name = "tinkerforge2mqtt"
package_version = "0.0.1"
package_description = "Emit MQTT events from tinkerforge"
package_url = "https://github.com/jedie/tinkerforge2mqtt"
issues_url = "https://github.com/jedie/tinkerforge2mqtt/issues"
license = "GPL-3.0-or-later"
_template = "https://github.com/jedie/cookiecutter_templates/"
